% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchResults.R
\name{catchActRaw}
\alias{catchActRaw}
\title{catchResults}
\usage{
catchActRaw(x)
}
\arguments{
\item{x}{The results from modeling fitting.}

\item{rawData}{Input raw data (quantitative phospho-proteomics / ubiqutin-proteomics data)}
}
\description{
Catch Activity and Affinity. Activity contains Raw-Activity, Mean-Center, Z-scaled.
}
