% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_referent.R
\name{create_referent}
\alias{create_referent}
\title{Create a referent list}
\usage{
create_referent(esem_efa_results)
}
\arguments{
\item{esem_efa}{is a \cite{psych::fa()} object with the results of exploratory factor analysis (EFA)
The object can be created using psych::fa() or a wrapper esem_efa() function
The function uses efa object to identify referents
A referent indicator is selected for each factor
It is the item that has a large (target) loading for the factor it measures and
The referents are used to ensure model identification and are used as starting values/ fixed values in the
the next step to create a lavaan model syntax.}
}
\value{
A list with factors and corresponding referents (i.e. referents in that factor)
}
\description{
Create a referent list
}
\examples{

# use Holzinger and Swineford (1939) dataset in lavaan package
hw_data <- lavaan::HolzingerSwineford1939
hw_data <- hw_data[,c(7:15)]

#make exploratory analysis with geomin rotation
esem_efa_results <- esem_efa(hw_data,3)
referent_list <- create_referent(esem_efa_results)
}
