% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_cfa.R
\name{esem_cfa}
\alias{esem_cfa}
\title{Confirmatory factor anaysis (CFA) step for ESEM-with-CFA}
\usage{
esem_cfa(model, data, mimic = c("MPlus"), std.lv = TRUE, ordered = TRUE)
}
\arguments{
\item{model}{is a character vector with a lavaan syntax for the ESEM model.}

\item{ordered}{}
}
\value{

}
\description{
is a wrapper for lavaan::cfa() function
}
