% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_cfa2.R
\name{esem_cfa2}
\alias{esem_cfa2}
\title{Exploratory Structural Equiation Modeling ESEM (ESEM)}
\usage{
esem_cfa2(
  data,
  nfactors,
  fm = "ML",
  rotate = "geominT",
  scores = "regression",
  residuals = TRUE,
  Target = NULL,
  missing = TRUE,
  mimic = c("MPlus"),
  std.lv = TRUE,
  ordered = TRUE
)
}
\arguments{
\item{ordered}{}
}
\value{

}
\description{
Exploratory Structural Equiation Modeling ESEM (ESEM)
}
