% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_cfa_syntax.R, R/esem_efa_syntax.R
\name{esem_cfa_syntax}
\alias{esem_cfa_syntax}
\title{Title}
\usage{
esem_cfa_syntax(loadings)

esem_cfa_syntax(loadings)
}
\arguments{
\item{esem_efa}{}
}
\value{



}
\description{
Title

Title
}
\examples{


}
