% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_efa.R
\name{esem_efa}
\alias{esem_efa}
\title{Exploratory factor analysis (EFA) for ESEM}
\usage{
esem_efa(
  data,
  nfactors,
  fm = "ML",
  rotate = "geominT",
  scores = "regression",
  residuals = TRUE,
  missing = TRUE
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Exploratory factor analysis (EFA) for ESEM
}
