% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_syntax_keys.R
\name{esem_syntax_keys}
\alias{esem_syntax_keys}
\title{Title}
\usage{
esem_syntax_keys(key_matrix, fixed)
}
\arguments{
\item{key_matrix}{is a matrix to be used to generate loadings}

\item{fixed}{is the values fixed or should be freely estimated}
}
\value{
A character vector with the syntax of the model tp be used at the CFA stage.
}
\description{
Title
}
