% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_syntax_mplus.R
\name{esem_syntax_mplus}
\alias{esem_syntax_mplus}
\title{Create a model syntax for ESEM-with-CFA compatible with MPlus}
\usage{
esem_syntax_mplus(key_matrix = NULL)
}
\arguments{
\item{key_matrix}{is a key matrix with the primary factor items. It can be made with the make.keys() function.
The primary factor items in the matrix are  used as referent items.}
}
\value{
A character vector with a lavaan syntax for the ESEM model that imitates MPlus.
}
\description{
Create a model syntax for ESEM-with-CFA compatible with MPlus
}
