% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa}
\alias{fa}
\title{Title}
\usage{
fa(
  data,
  nfactors = 2,
  rotate = "geominT",
  scores = "regression",
  residuals = TRUE,
  missing = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
