% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMEinput.R
\name{getFa}
\alias{getFa}
\title{Build MEME input}
\usage{
getFa(ks = NULL, filePath)
}
\arguments{
\item{ks}{Vector; Kinase species; ("mouse", "rat", "human")}

\item{filePath}{The path is used to save .fa files. Recommend to create a new folder to save all .fa.files.}
}
\description{
script to extract MEME input .fa file
}
\examples{
getFa(ks = "human", filePath = getwd())






}
