% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputTransform.R
\name{inputTransformPSP}
\alias{inputTransformPSP}
\title{inputTransform}
\usage{
inputTransformPSP(
  input_raw,
  input_log2_trans = FALSE,
  whole_log2_trans = FALSE,
  whole_proteome = NULL
)
}
\arguments{
\item{input_raw}{Input raw data (quantitative phospho-proteomics/ubiquitin-proteomics data)}

\item{input_log2_trans}{FALSE or TRUE. Need program to do log2 transforming of the input file or not. Default is FALSE.}

\item{whole_log2_trans}{Need program to do log2 transforming of the whole proteome or not. Ignore if -whole.proteome is missing. Default is FALSE.}

\item{whole_proteome}{Set up whole proteome used to normalize phospho-proteomics/ubiquitin-proteomics data by whole proteome. Default is NULL.}
}
\description{
This script is used to log2 transform AND/OR whole proteome normalize input data.
}
