% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanPlot.R
\name{lavaanPlot}
\alias{lavaanPlot}
\title{Plot the ESEM model}
\usage{
lavaanPlot(name = "plot", labels = NULL, ...)
}
\arguments{
\item{name}{A string of the name of the plot.}

\item{labels}{An optional named list of variable labels.}

\item{...}{Additional parameters as in lavaanPlot()}
}
\value{
A DiagrammeR plot of the path diagram for model
}
\description{
This is a wrapper function for lavaanPlot() from lavaanPlot() package.
For more information see https://cran.r-project.org/web/packages/lavaanPlot/index.html
}
\examples{
library(lavaan)
model <- 'mpg ~ cyl + disp + hp
          qsec ~ disp + hp + wt'
fit <- sem(model, data = mtcars)
lavaanPlot(model = fit, node_options = list(shape = "box", fontname = "Helvetica"),
          edge_options = list(color = "grey"), coefs = FALSE)
}
