% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_target.R
\name{make_target}
\alias{make_target}
\title{Title}
\usage{
make_target(data, keys)
}
\arguments{
\item{data}{is a dataset to be used in EFA}

\item{keys}{is a key matrix with the primary factor items. It can be made with the make.keys() function.
The primary factor items in the matrix are  used as referent items.}
}
\value{
a list with target matrix
}
\description{
Title
}
