% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastInput.R
\name{mastInput}
\alias{mastInput}
\title{Build MAST input}
\usage{
mastInput(input, savePath = getwd())
}
\arguments{
\item{input}{The same quantitative phosphoproteomics data as eSEM() input file.}

\item{savePath}{The path to save built MAST input file. Default is the current directory.}
}
\description{
Script to extract MAST input FASTA file.
}
\examples{
mastInput(input = input_psp_example)


}
