% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifEx.R
\name{motifEx}
\alias{motifEx}
\title{motifEx}
\usage{
motifEx(mastFile, savePath = getwd())
}
\arguments{
\item{mastFile}{The output file "extract_result.txt" from Step: 03step_ExtractMASTresult.sh.}

\item{savePath}{Path to save motif_input.txt file. Default is current directory.}
}
\description{
This script is used to extract kinase-substrate relationships from motif scanning and build motif_input file for KSEM().
}
