% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspAdjacency.R
\name{pspAdjacency}
\alias{pspAdjacency}
\title{pspAdjacency}
\usage{
pspAdjacency(sp)
}
\arguments{
\item{sp}{species; ("mouse", "rat", "human").}
}
\description{
This script is used to build kinase-substrate adjacency matrix with only PSP database.
}
