% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleHatAct.R
\name{singleHatAct}
\alias{singleHatAct}
\title{singleHatAct}
\usage{
singleHatAct(HAT, input, adj, cor.off, kmo.off)
}
\arguments{
\item{HAT}{Single HAT name.}

\item{input}{Normalized and transformed acetyl-proteomics data.}

\item{adj}{Adjacency matrix.}

\item{cor.off}{Set up correlation cutoff value 0-1 to remove high collinear variables. Default is 0.95.}

\item{kmo.off}{Set up KMO cutoff value 0-1. Default is 0.}
}
\description{
Get single HAT activity.
}
