% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiAdjacency.R
\name{ubiAdjacency}
\alias{ubiAdjacency}
\title{ubiAdjacency}
\usage{
ubiAdjacency(sp)
}
\arguments{
\item{sp}{Substrate species; ("mouse", "rat", "human").}
}
\description{
This script is used to build ligase-substrate adjacency matrix with only PSP database.
}
