% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_referent.R
\name{create_referent}
\alias{create_referent}
\title{Create a referent list}
\usage{
create_referent(esem_efa_results)
}
\arguments{
\item{esem_efa_results}{is a \cite{psych::fa()} object with the results of exploratory factor analysis (EFA)
The object can be created using psych::fa() or a wrapper esem_efa() function
The function uses efa object to identify referents
A referent indicator is selected for each factor
It is the item that has a large (target) loading for the factor it measures and
The referents are used to ensure model identification and are used as starting values/ fixed values in the
the next step to create a lavaan model syntax.}
}
\value{
A list with factors and corresponding referents (i.e. referents in that factor)
}
\description{
Create a referent list
}
