% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem.R
\name{esem}
\alias{esem}
\title{esem
completes esem estimation using the provided parameters}
\usage{
esem(awesome, beautiful)
}
\arguments{
\item{awesome}{is the parameter to control}

\item{beautiful}{is the parameter to control}
}
\value{
prints the string
}
\description{
esem
completes esem estimation using the provided parameters
}
\examples{
esem("awesome", "beautiful")
}
