% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSEM.R
\name{eSEM}
\alias{eSEM}
\title{eSEM}
\usage{
eSEM(
  input = NULL,
  datatype = NULL,
  organism = NULL,
  enzyme.organism = NULL,
  cor.off = 0.95,
  kmo.off = 0,
  enzyList = NULL,
  input.log2.norm = FALSE,
  whole.log2.trans = FALSE,
  whole.proteome = NULL,
  motif = NULL,
  output.folder = getwd()
)
}
\arguments{
\item{input}{Dataframe. Your quantitative ubiquiti-proteomics, phospho-proteomic and acetyl-proteomics data.}

\item{datatype}{Factor. Data type: ubiquitination or phosphoralation or acetylation; ("ubi", "psp", "ace")}

\item{organism}{Vector. Substrate species; (“human”, “mouse”, “rat”)}

\item{enzyme.organism}{Vector. Enzyme species; (“human”, “mouse”, “rat”)}

\item{cor.off}{Numeric. Set up correlation cutoff value 0-1 to remove high collinear variables. Default is 0.95.}

\item{kmo.off}{Numeric. Set up KMO cutoff value 0-1. Default is 0.}

\item{enzyList}{Vector. Program only calculate the enzyme in the enzyList. Default is to output ALL enzyme activities and affinities.}

\item{input.log2.norm}{Logical. FALSE or TRUE. Need program to do log2 transforming of the input file or not. Default is FALSE.}

\item{whole.log2.trans}{Logical. FALSE or TRUE. Need program to do log2 transforming of the whole proteome or not. (Ignore if -whole.proteome is missing). Default is FALSE.}

\item{whole.proteome}{Dataframe. Set up whole proteome used to normalize phosphor-proteome or ubiquitin-proteome whole proteome. Default is NULL.}

\item{motif}{Matrix. Added kinase-subatrate relationships from motif discovery.}

\item{output.folder}{Character. Character vector of location to save files if desired. Default is current directory.}
}
\value{
List
}
\description{
Infer enzyme activity from phosphoralation, ubiquitination, acetylation
}
\details{
You can use eSEM package to get enzyme activity.
}
\examples{

result <- eSEM(input = input_ubi_example,
               datatype = "ubi",
               organism = "human",
               input.log2.norm = T)


result <- eSEM(input = input_psp_example,
               datatype = "psp",
               organism = "mouse",
               input.log2.norm = TRUE,
               whole.log2.trans = TRUE,
               motif = motif_example,
               whole.proteome = wholeProteome_example)


result <- eSEM(input = input_ace_example,
               datatype = "ace",
               organism = "human",
               input.log2.norm = FALSE)




}
