% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esem_cfa.R
\name{esem_cfa}
\alias{esem_cfa}
\title{Confirmatory factor anaysis (CFA) step for ESEM-with-CFA}
\usage{
esem_cfa(model, data, mimic = c("MPlus"), std.lv = TRUE, ordered = TRUE)
}
\arguments{
\item{model}{is a character vector with a lavaan syntax for the ESEM model.}

\item{data}{is a raw data matrix.}

\item{mimic}{allows to mimic the final output results (i.e. CFA stage) to MPLUS to allow the user to compare the output.}

\item{std.lv}{is set to TRUE by default to provide standardized latent variables.}

\item{ordered}{is set to TRUE by default to allow the use of categorical variables.}
}
\value{
An object of class \cite{lavaan::lavaan-class}, for which several methods are available, including a summary method.
}
\description{
is a wrapper for lavaan::cfa() function
}
