% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMEinput.R
\name{getFa}
\alias{getFa}
\title{Build MEME input}
\usage{
getFa(ss = NULL, es = NULL, filePath)
}
\arguments{
\item{ss}{Vector; Substrate species; ("mouse", "rat", "human")}

\item{es}{Vector; Kinase species; ("mouse", "rat", "human")}

\item{filePath}{The path is used to save .fa files. Recommend to create a new folder to save all .fa.files.}
}
\description{
script to extract MEME input .fa file
}
\examples{
getFa(ss = "human", es = c("human", "mouse", "rat"), filePath = getwd())






}
