% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiAdjacency.R
\name{ubiAdjacency}
\alias{ubiAdjacency}
\title{ubiAdjacency}
\usage{
ubiAdjacency(sp, databaseUBI)
}
\arguments{
\item{sp}{Substrate species; ("mouse", "rat", "human").}

\item{databaseUBI}{Default database or customized database input.}
}
\description{
This script is used to build ligase-substrate adjacency matrix with only PSP database.
}
